Added support for Kdyby/Console
